backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async HTTP Client
httpx==0.26.0
requests==2.31.0

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Firebase
firebase-admin==6.4.0

# AI Services
openai==1.12.0
replicate==0.25.1

# PDF Generation
reportlab==4.0.9

# Image Processing
numpy>=1.24.0
Pillow>=10.0.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
stripe==14.1.0

# Email (falls wir resend nutzen, sicherheitshalber hinzufügen)
resend==0.8.0
